test: move NetIfaceConfig to network.py
Move it there so it is with other networking classes.
