+++ tests/host_tools/network.py
@@ -4,6 +4,7 @@
 
 import random
 import string
+from dataclasses import dataclass
 from pathlib import Path
 
 from tenacity import retry, retry_if_exception_type, stop_after_attempt, wait_fixed
@@ -182,3 +183,29 @@ class Tap:
 
     def __repr__(self):
         return f"<Tap name={self.name} netns={self.netns}>"
+
+
+@dataclass(frozen=True, repr=True)
+class NetIfaceConfig:
+    """Defines a network interface configuration."""
+
+    host_ip: str = "192.168.0.1"
+    guest_ip: str = "192.168.0.2"
+    tap_name: str = "tap0"
+    dev_name: str = "eth0"
+    netmask: int = 30
+
+    @property
+    def guest_mac(self):
+        """Return the guest MAC address."""
+        return mac_from_ip(self.guest_ip)
+
+    @staticmethod
+    def with_id(i):
+        """Define network iface with id `i`."""
+        return NetIfaceConfig(
+            host_ip=f"192.168.{i}.1",
+            guest_ip=f"192.168.{i}.2",
+            tap_name=f"tap{i}",
+            dev_name=f"eth{i}",
+        )
